refactor(client): dedupe chat panel controls, copy-markdown, drop overlay
Extract three verbatim duplicates shared by the docked and floating chat
surfaces, no behavior change:

- useCopySessionMarkdown: the copy-chat-as-markdown callback + copied
  flag, previously duplicated in DockedChatPanel and FloatingChatWindow
- ChatPanelControls: the copy/float/dock-right/dock-bottom button
  cluster, parameterized by testid prefix and active dock layout
- DropFilesOverlay: the drag-over blur overlay repeated three times in
  SessionContainer
