+++ apps/client/app/components/Session/DockedChatPanel.tsx
@@ -1,18 +1,10 @@
 'use client';
 
-import React, { useCallback, useState } from 'react';
+import React, { useCallback } from 'react';
 import { Box, IconButton, Typography, Tooltip } from '@mui/joy';
-import OpenInNewIcon from '@mui/icons-material/OpenInNew';
-import VerticalSplitIcon from '@mui/icons-material/VerticalSplit';
-import HorizontalSplitIcon from '@mui/icons-material/HorizontalSplit';
 import CloseIcon from '@mui/icons-material/Close';
-import ContentCopyIcon from '@mui/icons-material/ContentCopy';
-import CheckIcon from '@mui/icons-material/Check';
 import useSessionLayout, { setSessionLayout } from '@client/app/hooks/useSessionLayout';
-import { useSessions } from '@client/app/contexts/SessionsContext';
-import { useQueryClient, type InfiniteData } from '@tanstack/react-query';
-import type { IChatHistoryItemDocument } from '@bike4mind/common';
-import { convertSessionToMarkdown } from '@client/app/utils/sessionMarkdownExport';
+import ChatPanelControls from './ChatPanelControls';
 
 interface DockedChatPanelProps {
   children: React.ReactNode;
@@ -20,43 +12,7 @@ interface DockedChatPanelProps {
 }
 
 const DockedChatPanel: React.FC<DockedChatPanelProps> = ({ children, headerActions }) => {
-  const [copied, setCopied] = useState(false);
   const layout = useSessionLayout(s => s.layout);
-  const { currentSessionId } = useSessions();
-  const queryClient = useQueryClient();
-
-  const handleCopyMarkdown = useCallback(async () => {
-    if (!currentSessionId) return;
-    try {
-      const queryData = queryClient.getQueryData<InfiniteData<{ data: IChatHistoryItemDocument[] }>>([
-        'quests',
-        'session',
-        currentSessionId,
-      ]);
-      if (!queryData?.pages) return;
-
-      const quests = queryData.pages.flatMap(p => p.data).reverse();
-      const markdown = convertSessionToMarkdown(quests);
-
-      await navigator.clipboard.writeText(markdown);
-      setCopied(true);
-      setTimeout(() => setCopied(false), 2000);
-    } catch (err) {
-      console.error('Failed to copy markdown:', err);
-    }
-  }, [currentSessionId, queryClient]);
-
-  const handleSwitchToFloat = useCallback(() => {
-    setSessionLayout({ layout: 'floatingChat', previousLayout: layout });
-  }, [layout]);
-
-  const handleSwitchToDockRight = useCallback(() => {
-    setSessionLayout({ layout: 'dockRight' });
-  }, []);
-
-  const handleSwitchToDockBottom = useCallback(() => {
-    setSessionLayout({ layout: 'dockBottom' });
-  }, []);
 
   const handleClose = useCallback(() => {
     setSessionLayout({ layout: 'floatingChat' });
@@ -102,54 +58,11 @@ const DockedChatPanel: React.FC<DockedChatPanelProps> = ({ children, headerActio
           {/* Session actions first (headerActions leads with the primary New Chat),
               then window controls. */}
           {headerActions}
-          <Tooltip title={copied ? 'Copied!' : 'Copy chat as Markdown'} disableInteractive>
-            <IconButton
-              size="sm"
-              variant="plain"
-              color={copied ? 'success' : 'neutral'}
-              onClick={handleCopyMarkdown}
-              data-testid="docked-chat-copy-markdown"
-              sx={{ '--IconButton-size': '28px' }}
-            >
-              {copied ? <CheckIcon sx={{ fontSize: 16 }} /> : <ContentCopyIcon sx={{ fontSize: 14 }} />}
-            </IconButton>
-          </Tooltip>
-          <Tooltip title="Float" disableInteractive>
-            <IconButton
-              size="sm"
-              variant="plain"
-              color="neutral"
-              onClick={handleSwitchToFloat}
-              data-testid="docked-chat-float"
-              sx={{ '--IconButton-size': '28px' }}
-            >
-              <OpenInNewIcon sx={{ fontSize: 16 }} />
-            </IconButton>
-          </Tooltip>
-          <Tooltip title="Dock right" disableInteractive>
-            <IconButton
-              size="sm"
-              variant={layout === 'dockRight' ? 'soft' : 'plain'}
-              color={layout === 'dockRight' ? 'primary' : 'neutral'}
-              onClick={handleSwitchToDockRight}
-              data-testid="docked-chat-dock-right"
-              sx={{ '--IconButton-size': '28px' }}
-            >
-              <VerticalSplitIcon sx={{ fontSize: 16 }} />
-            </IconButton>
-          </Tooltip>
-          <Tooltip title="Dock bottom" disableInteractive>
-            <IconButton
-              size="sm"
-              variant={layout === 'dockBottom' ? 'soft' : 'plain'}
-              color={layout === 'dockBottom' ? 'primary' : 'neutral'}
-              onClick={handleSwitchToDockBottom}
-              data-testid="docked-chat-dock-bottom"
-              sx={{ '--IconButton-size': '28px' }}
-            >
-              <HorizontalSplitIcon sx={{ fontSize: 16 }} />
-            </IconButton>
-          </Tooltip>
+          <ChatPanelControls
+            testIdPrefix="docked-chat"
+            activeLayout={layout === 'dockRight' || layout === 'dockBottom' ? layout : undefined}
+            showFloat
+          />
           <Tooltip title="Close" disableInteractive>
             <IconButton
               size="sm"
