+++ apps/client/app/components/Session/FloatingChatWindow.tsx
@@ -9,15 +9,8 @@ import DragIndicatorIcon from '@mui/icons-material/DragIndicator';
 import CloseIcon from '@mui/icons-material/Close';
 import MinimizeIcon from '@mui/icons-material/Minimize';
 import SmartToyIcon from '@mui/icons-material/SmartToy';
-import ContentCopyIcon from '@mui/icons-material/ContentCopy';
-import CheckIcon from '@mui/icons-material/Check';
-import VerticalSplitIcon from '@mui/icons-material/VerticalSplit';
-import HorizontalSplitIcon from '@mui/icons-material/HorizontalSplit';
 import useSessionLayout, { setSessionLayout } from '@client/app/hooks/useSessionLayout';
-import { useSessions } from '@client/app/contexts/SessionsContext';
-import { useQueryClient, type InfiniteData } from '@tanstack/react-query';
-import type { IChatHistoryItemDocument } from '@bike4mind/common';
-import { convertSessionToMarkdown } from '@client/app/utils/sessionMarkdownExport';
+import ChatPanelControls from './ChatPanelControls';
 
 // react-draggable v4.7 types DraggableProps to cover nodeRef/handle/position/
 // bounds/onDrag/onStop directly, so no cast is needed.
@@ -91,9 +84,6 @@ const getCursorForHandle = (position: string): string => {
 const FloatingChatWindow: React.FC<FloatingChatWindowProps> = ({ children, headerActions }) => {
   const nodeRef = useRef<HTMLDivElement>(null);
   const [isResizing, setIsResizing] = useState(false);
-  const [copied, setCopied] = useState(false);
-  const { currentSessionId } = useSessions();
-  const queryClient = useQueryClient();
 
   // Get state from Zustand store
   const position = useSessionLayout(s => s.floatingChatPosition);
@@ -113,28 +103,6 @@ const FloatingChatWindow: React.FC<FloatingChatWindowProps> = ({ children, heade
     el.addEventListener('wheel', handleWheel, { passive: true });
     return () => el.removeEventListener('wheel', handleWheel);
   }, [isMinimized]);
-
-  // Copy entire chat history as markdown to clipboard
-  const handleCopyMarkdown = useCallback(async () => {
-    if (!currentSessionId) return;
-    try {
-      const queryData = queryClient.getQueryData<InfiniteData<{ data: IChatHistoryItemDocument[] }>>([
-        'quests',
-        'session',
-        currentSessionId,
-      ]);
-      if (!queryData?.pages) return;
-
-      const quests = queryData.pages.flatMap(p => p.data).reverse(); // chronological order
-      const markdown = convertSessionToMarkdown(quests);
-
-      await navigator.clipboard.writeText(markdown);
-      setCopied(true);
-      setTimeout(() => setCopied(false), 2000);
-    } catch (err) {
-      console.error('Failed to copy markdown:', err);
-    }
-  }, [currentSessionId, queryClient]);
 
   // Local state for drag/resize operations (for smooth updates)
   const [localPosition, setLocalPosition] = useState(position);
@@ -456,42 +424,7 @@ const FloatingChatWindow: React.FC<FloatingChatWindowProps> = ({ children, heade
             onPointerDown={e => e.stopPropagation()}
           >
             {headerActions}
-            <Tooltip title={copied ? 'Copied!' : 'Copy chat as Markdown'} disableInteractive>
-              <IconButton
-                size="sm"
-                variant="plain"
-                color={copied ? 'success' : 'neutral'}
-                onClick={handleCopyMarkdown}
-                data-testid="floating-chat-copy-markdown"
-                sx={{ '--IconButton-size': '28px' }}
-              >
-                {copied ? <CheckIcon sx={{ fontSize: 16 }} /> : <ContentCopyIcon sx={{ fontSize: 14 }} />}
-              </IconButton>
-            </Tooltip>
-            <Tooltip title="Dock right" disableInteractive>
-              <IconButton
-                size="sm"
-                variant="plain"
-                color="neutral"
-                onClick={() => setSessionLayout({ layout: 'dockRight' })}
-                data-testid="floating-chat-dock-right"
-                sx={{ '--IconButton-size': '28px' }}
-              >
-                <VerticalSplitIcon sx={{ fontSize: 16 }} />
-              </IconButton>
-            </Tooltip>
-            <Tooltip title="Dock bottom" disableInteractive>
-              <IconButton
-                size="sm"
-                variant="plain"
-                color="neutral"
-                onClick={() => setSessionLayout({ layout: 'dockBottom' })}
-                data-testid="floating-chat-dock-bottom"
-                sx={{ '--IconButton-size': '28px' }}
-              >
-                <HorizontalSplitIcon sx={{ fontSize: 16 }} />
-              </IconButton>
-            </Tooltip>
+            <ChatPanelControls testIdPrefix="floating-chat" />
             <Tooltip title="Minimize" disableInteractive>
               <IconButton
                 size="sm"
